api/requirements.txt
Flask
flask_restful
gunicorn
Werkzeug
python-dotenv
flask_cors
requests